Move getInitials into utils/common.ts
Fold the helper and its tests into the shared common module instead of a
single-function file; update the InitialsAvatar import.
